Introduce the concept of annotations queriable at run time
Added the getter get_script_property_annotations to the Script class.
This will allow each scripting language to implement the concept
of annotations if they wish.

The idea is to enable users to mark their script members with
static metadata queriable at run time. This is excepected to be useful
for plug-ins to provide special handling for annotated members.
For example, a plug in could create a custom inspector widget
that is enabled and condfigured by a specific annotation.