Implement annotation thunks to lazily instantiate real annotations
The annotations can't be instantiated at compile time,
as creating live instances of scripts is not possible then.
So instead we store the data for later.

When somebody asks for the annotations, we instantiate them for real.
We need to load the script resource and set it as the script
to make this work. We'll also need a separate branch
for the built-in classes later, where we use ClassDB::instantiate.
I think. This is base on the code in editor_data.cpp TOOL_CREATE_FAVORITE.